Add files via upload
This project aims to offer a simple and effective method for sorting and visualizing data. Use: Any computer with the Java runtime environment installed can use this. Most frequently applied in educational and scientific contexts. Function: Stores data from a file in an array or an ArrayList data structure after reading it from the file.